+++ cmsdev/internal/test/bos/bos.go
@@ -53,7 +53,7 @@ func getAnyTenant(tenantList []string) string {
 	return defaultTenantName
 }
 
-func IsBOSRunning(includeCLI bool) (passed bool) {
+func IsBOSRunning(includeCLI, includeTenant bool) (passed bool) {
 	var err error
 	var tenantList []string
 	artifactsCollected := false
@@ -120,14 +120,14 @@ func IsBOSRunning(includeCLI bool) (passed bool) {
 	}
 
 	// Defined in bos_api.go
-	if !apiTests(tenantList) {
+	if !apiTests(tenantList, includeTenant) {
 		passed = false
 	}
 
 	// CLI tests will be run only if requested using the include-cli flag
 	if includeCLI {
 		// Defined in bos_cli.go
-		if !cliTests(tenantList) {
+		if !cliTests(tenantList, includeTenant) {
 			passed = false
 		}
 	}